+++ dart/test/sentry_exception_factory_test.dart
@@ -83,6 +83,28 @@ void main() {
 
     // skip on browser because [StackTrace.current] still returns null
   }, onPlatform: {'browser': Skip()});
+
+  test('reads the snapshot from the mechanism', () {
+    final error = StateError('test-error');
+    final mechanism = Mechanism(type: 'Mechanism');
+    final throwableMechanism = ThrowableMechanism(
+      mechanism,
+      error,
+      snapshot: true,
+    );
+
+    SentryException sentryException;
+    try {
+      throw throwableMechanism;
+    } catch (err, stackTrace) {
+      sentryException = fixture.getSut().getSentryException(
+            throwableMechanism,
+            stackTrace: stackTrace,
+          );
+    }
+
+    expect(sentryException.stackTrace!.snapshot, true);
+  });
 }
 
 class CustomError extends Error {}