+++ flutter/lib/src/integrations/on_error_integration.dart
@@ -56,6 +56,8 @@ class OnErrorIntegration implements Integration<SentryFlutterOptions> {
       var event = SentryEvent(
         throwable: throwableMechanism,
         level: SentryLevel.fatal,
+        // ignore: invalid_use_of_internal_member
+        timestamp: options.clock(),
       );
 
       // unawaited future